fix: Refactoring from zero opacity to hide element
