+++ e2eTests/chrome/hideElements/baseline/hide-element-large.png
