+++ e2eTests/firefox/hideElements/baseline/hide-element-large.png
