+++ e2eTests/firefox/hideElements/hideElements.test.js
@@ -3,10 +3,9 @@
 import { execSync } from 'child_process';
 import path from 'path';
 import fs from 'fs';
-import config from './zeroOpacityElementsConfig';
+import config from './hideElementsConfig';
 
 function cleanState(dir) {
-  console.log('CLEANING STATE');
   if (fs.existsSync(dir)) {
     const files = fs.readdirSync(dir);
     files.forEach(file => fs.unlinkSync(`${dir}/${file}`));
@@ -14,25 +13,23 @@ function cleanState(dir) {
   }
 }
 
-describe('e2e Tests zero opacity elements', () => {
+describe('e2e Tests hide elements', () => {
   let latestMockImagesPath;
 
   beforeEach(() => {
     latestMockImagesPath = path.resolve(config.latest);
-
-    //cleanState(latestMockImagesPath);
   });
 
-  it('set zero opacity elements from webpage and compares', () => {
+  it('hides elements using zero opacity', () => {
     let exitCode = 0;
 
     try {
       const snapOutput = execSync(
-        'node ./lib/bin/run.js snap --browser firefox --config e2eTests/firefox/zeroOpacityElements/zeroOpacityElementsConfig.json'
+        'node ./lib/bin/run.js snap --browser firefox --config e2eTests/firefox/hideElements/hideElementsConfig.json'
       ).toString();
       console.log(snapOutput);
       const compareOutput = execSync(
-        'node ./lib/bin/run.js compare --browser firefox --config e2eTests/firefox/zeroOpacityElements/zeroOpacityElementsConfig.json'
+        'node ./lib/bin/run.js compare --browser firefox --config e2eTests/firefox/hideElements/hideElementsConfig.json'
       ).toString();
       console.log(compareOutput);
     } catch (err) {
