--- e2eTests/firefox/zeroOpacityElements/zeroOpacityElementsConfig.json
@@ -1,15 +1,0 @@
-{
-  "gridUrl": "http://hub:4444/wd/hub",
-  "baseline": "./e2eTests/firefox/zeroOpacityElements/baseline",
-  "latest": "./e2eTests/firefox/zeroOpacityElements/latest",
-  "generatedDiffs": "./e2eTests/firefox/zeroOpacityElements/generatedDiffs",
-  "report": "./e2eTests/firefox/zeroOpacityElements/reports",
-  "scenarios": [
-    {
-      "url": "http://ayespy_report:4000",
-      "label": "zero-opacity-element",
-      "viewports": [{"height": 2400, "width": 1024, "label": "large"}],
-      "zeroOpacityElements": [".screenshot"]
-    }
-  ]
-}
